Add custom authentication option.

+++ app/Http/Controllers/PostController.php
@@ -6,7 +6,19 @@ use App\Like;
 use App\Post;
 use App\Tag;
 use Auth;
+use Gate; // for custom gate
 use Illuminate\Http\Request;
+
+
+// General AUTH notes:
+// Try to authenticate user example
+// Auth::attempt(['email'=>$email, 'password'=>$password]);
+// Check if active user is authenticated
+// Auth::check();
+// Get authenticated user, returns null if not
+// Auth::user();
+// Log user out
+// Auth::logout();
 
 class PostController extends Controller
 {
@@ -20,6 +32,10 @@ class PostController extends Controller
 
 	public function getAdminIndex() 
 	{
+		if (!Auth::check()) {
+			return redirect()->back();
+		}
+
 		// $posts = Post::all();
 		// $posts = Post::orderBy('title', 'asc')->get();
 		// $posts = Post::orderBy('title', 'asc')->paginate(3);
@@ -46,12 +62,21 @@ class PostController extends Controller
 
 	public function getAdminCreate()
 	{
+		if (!Auth::check()) {
+			return redirect()->back();
+		}
+
 		$tags = Tag::all();
 		return view('admin.create', ['tags' => $tags]);
 	}
 
 	public function getAdminEdit($id)
 	{
+		
+		if (!Auth::check()) {
+			return redirect()->back();
+		}
+
 		$post = Post::find($id);
 		$tags = Tag::all();
 		return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
@@ -59,6 +84,10 @@ class PostController extends Controller
 
 	public function postAdminCreate(Request $request)
 	{
+		if (!Auth::check()) {
+			return redirect()->back();
+		}
+
 		$this->validate($request, [
 			'title' => 'required|min:5',
 			'content' => 'required|min:10'
@@ -83,11 +112,20 @@ class PostController extends Controller
 
 	public function postAdminUpdate(Request $request)
 	{
+		if (!Auth::check()) {
+			return redirect()->back();
+		}
+
 		$this->validate($request, [
 			'title' => 'required|min:5',
 			'content' => 'required|min:10'
 		]);
 		$post = Post::find($request->input('id'));
+
+		if (Gate::denies('manipulate-post', $post)) {
+			return redirect()->route('admin.index')->with('info', 'Update denied. You didn\'t write that shit.'); // if denied, redirect
+		}
+
 		$post->title = $request->input('title');
 		$post->content = $request->input('content');
 		$post->save();
@@ -102,7 +140,16 @@ class PostController extends Controller
 
 	public function getAdminDelete($id)
 	{
+		if (!Auth::check()) {
+			return redirect()->back();
+		}
+
 		$post = Post::find($id);
+
+		if (Gate::denies('manipulate-post', $post)) {
+			return redirect()->back()->with('info', 'Delete denied.'); // if denied, redirect
+		}
+
 		$post->likes()->delete();
 		$post->tags()->detach();
 		$post->delete();
